+++ test/Core/Exploration/ExplorerStackTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Core\Test\Exploration;
 
 use Heptacom\HeptaConnect\Core\Exploration\ExplorerStack;
-use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Core\Test\Fixture\FooBarEntity;
 use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExploreContextInterface;
 use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExplorerContract;
 use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExplorerStackInterface;
@@ -14,6 +14,11 @@ use Psr\Log\LoggerInterface;
 
 /**
  * @covers \Heptacom\HeptaConnect\Core\Exploration\ExplorerStack
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringReferenceContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\SubtypeClassStringContract
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\EntityType
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection
  * @covers \Heptacom\HeptaConnect\Portal\Base\Exploration\ExplorerCollection
@@ -22,7 +27,7 @@ final class ExplorerStackTest extends TestCase
 {
     public function testEmptyStackDoesNotFail(): void
     {
-        $stack = new ExplorerStack([], $this->createMock(LoggerInterface::class));
+        $stack = new ExplorerStack([], FooBarEntity::class(), $this->createMock(LoggerInterface::class));
         static::assertCount(0, $stack->next(
             $this->createMock(ExploreContextInterface::class)
         ));
@@ -30,9 +35,9 @@ final class ExplorerStackTest extends TestCase
 
     public function testStackCallsEveryone(): void
     {
-        $result1 = $this->createMock(DatasetEntityContract::class);
-        $result2 = $this->createMock(DatasetEntityContract::class);
-        $result3 = $this->createMock(DatasetEntityContract::class);
+        $result1 = new FooBarEntity();
+        $result2 = new FooBarEntity();
+        $result3 = new FooBarEntity();
 
         $explorer1 = $this->createMock(ExplorerContract::class);
         $explorer1->expects(static::once())
@@ -57,7 +62,11 @@ final class ExplorerStackTest extends TestCase
                 yield from $stack->next($c);
             });
 
-        $stack = new ExplorerStack([$explorer1, $explorer2, $explorer3], $this->createMock(LoggerInterface::class));
+        $stack = new ExplorerStack(
+            [$explorer1, $explorer2, $explorer3],
+            FooBarEntity::class(),
+            $this->createMock(LoggerInterface::class)
+        );
         static::assertCount(3, $stack->next($this->createMock(ExploreContextInterface::class)));
     }
 }